--- a/public/app/plugins/datasource/loki/shardQuerySplitting.test.ts
+++ b/public/app/plugins/datasource/loki/shardQuerySplitting.test.ts
@@ -79,7 +79,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_0_2',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_0_2_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"20|10"}',
@@ -92,7 +93,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_2_2',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_2_2_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"3|2"}',
@@ -105,7 +107,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_4_1',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_4_1_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__="1"}',
@@ -118,7 +121,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_5_1',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_5_1_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=""}',
@@ -147,7 +151,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_0_2',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_0_2_1',
         targets: [
           {
             expr: '{service_name="test", filter="true", __stream_shard__=~"20|10"}',
@@ -208,18 +213,7 @@ describe('runShardSplitQuery()', () => {
   });
 
   test('Adjusts the group size based on errors and execution time', async () => {
-    const request = createRequest([{ expr: '$SELECTOR', refId: 'A', direction: LokiQueryDirection.Scan }], {
-      range: {
-        from: dateTime('2024-11-13T05:00:00.000Z'),
-        to: dateTime('2024-11-14T06:00:00.000Z'),
-        raw: {
-          from: dateTime('2024-11-13T05:00:00.000Z'),
-          to: dateTime('2024-11-14T06:00:00.000Z'),
-        },
-      },
-    });
-
-    jest
+   jest
       .mocked(datasource.languageProvider.fetchLabelValues)
       .mockResolvedValue(['1', '10', '2', '20', '3', '4', '5', '6', '7', '8', '9']);
 
@@ -373,7 +367,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_0_3',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_0_3_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"20|10|9"}',
@@ -387,7 +382,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_3_4',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_3_4_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"8|7|6|5"}',
@@ -401,7 +397,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_3_2',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_3_2_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"8|7"}',
@@ -415,7 +412,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_5_3',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_5_3_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"6|5|4"}',
@@ -429,7 +427,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_8_2',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_8_2_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=~"3|2"}',
@@ -443,7 +442,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_10_1',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_10_1_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__="1"}',
@@ -457,7 +457,8 @@ describe('runShardSplitQuery()', () => {
       expect(datasource.runQuery).toHaveBeenCalledWith({
         intervalMs: expect.any(Number),
         range: expect.any(Object),
-        requestId: 'TEST_shard_0_11_1',
+        queryGroupId: expect.any(String),
+        requestId: 'TEST_shard_0_11_1_1',
         targets: [
           {
             expr: '{a="b", __stream_shard__=""}',