This PR updates shard query splitting to execute queries through the time query splitting runner. This allows multi-day queries to also benefit from time splitting, improving queries in situations with very high volume (huge amount of logs over several days), where the lower shards (0, 1) or the empty shard queries will timeout and never execute.