@matyax 2025-01-14
@grafana/observability-logs mind looking at this one? It should further improve Scan users for multi-day queries.