@paulojmdias 2025-01-20
@matyax were tested with large datasets?

We are seeing a performance drop using this feature.

I will let here some time details.

- Without this FF enabled - Using version 11.4.0
  - 12h query - 8s - This query will process approximately 520GiB.
  - 24h query - 10s - This query will process approximately 1.1 TiB.
  - 48h query - 18s - This query will process approximately 1.8 TiB.
- With this FF enabled - Using the main version
  - 12h query - 24s - This query will process approximately 520GiB.
  - 24h query - 24s - This query will process approximately 1.1 TiB.
  - 48h query - 26s - This query will process approximately 1.8 TiB.

We only see performance improvements if we query more than 4 days. On all the tests above, Grafana chooses the 1m step. Only when we test 4 days, Grafana switches the step to 1h.