@matyax 2025-01-20
@paulojmdias 

> sorry for the comment on a closed PR, if needed I can open an issue

Not a problem at all!

> We are seeing a performance drop using this feature. 

There's a lot to unwrap here. It would be helpful to know what queries you are running (metric or logs, and some examples) and in which environments these tests are being done.

For example, Shard splitting for logs is used only when the query direction is `Scan`, and should only be used for needle-in-a-haystack queries looking for something specific across a big data set.

For metric queries, Shard splitting aims at getting a first result faster by querying lighter shards first. Also I'm not sure if you mean only shard splitting or this change that combines shard and time splitting.

In any case, we look forward to getting feedback like this to implement better querying strategies. If you don't mind, I'd appreciate it if you can open an issue with these answers (and links to OPS or other instances) so we can use that for any further discussion.

And, by the way, to help benchmarking queries you can disable sharding in Explore by changing the query refId to "do-not-shard", like this:

![imagen](https://github.com/user-attachments/assets/4c30c7a6-720f-4516-99b7-c38b450203a1)