@paulojmdias 2025-01-20
Thanks for the quick answer @matyax 

> There's a lot to unwrap here. It would be helpful to know what queries you are running (metric or logs, and some examples) and in which environments these tests are being done.

The only measure we are doing is the time needed for the Logs Volume panel to be completely loaded. And with the changes related to `lokiShardSplitting` feature flag, seems slower in the short time ranges.

<img width="3183" alt="image" src="https://github.com/user-attachments/assets/8cf69340-8049-492b-b050-8bbe831877b4" />

The query used is something like:

```
{provider="google", service="service1", role="role1",} | instance=`instance1.domain1.com` | json
```

> In any case, we look forward to getting feedback like this to implement better querying strategies. If you don't mind, I'd appreciate it if you can open an issue with these answers (and links to OPS or other instances) so we can use that for any further discussion.

We are testing this using open source on our internal deployment, do you see value in opening an issue to collaborate on this? Technically I think it was also related to the closed one https://github.com/grafana/grafana/issues/68956