@matyax 2025-01-21
> The only measure we are doing is the time needed for the Logs Volume panel to be completely loaded.

Great to know.

> We are testing this using open source on our internal deployment, do you see value in opening an issue to collaborate on this? 

Yes, absolutely. We'd use the issue to have a dedicated discussion and track changes, if any.

> Technically I think it was also related to the closed one https://github.com/grafana/grafana/issues/68956

The main inconvenient with Logs Volume was timeouts, not total execution time.

---

For now, my recommendation would be to only enable `lokiShardSplitting` if 1) [sharding](https://grafana.com/docs/loki/latest/operations/automatic-stream-sharding/) is enabled and 2) you're experiencing timeouts or would prefer it drawn with partial results over the whole range, like:

![2025-01-21 10 55 06](https://github.com/user-attachments/assets/d447596d-a338-41ae-ab54-ca243edfec6f)

If not, it would be then recommended to stick to time splitting, which is the default behavior.